// src/models/UtPcaFormId.ts
import { UtPcaContProdId } from './UtPcaContProdId';
import { UtPcaFormDefnId } from './UtPcaFormDefnId';
export class UtPcaFormId {
	formId?: number;
	contProdId?: number;
	deletedFlag?: string;
	formNm?: string;
	formCritCntId?: number;
	formTypeAId?: number;
	formTypeBId?: number;
	formTypeCId?: number;
	mktbsktId?: number;
	mktbsktId1?: number;
	mktbsktId2?: number;
	mktbsktId3?: number;
	mktbsktId4?: number;
	mktbsktId5?: number;
	mktbsktId6?: number;
	mktbsktId7?: number;
	mktbsktId8?: number;
	formNote?: string;
	clinicalEditEval?: boolean;
	formCreateDate?: string;
	formUpdateDate?: string;
	formTypeAId1?: number;
	formTypeAId2?: number;
	formTypeAId3?: number;
	formTypeAId4?: number;
	formTypeAId5?: number;
	formTypeAId6?: number;
	formTypeAId7?: number;
	formTypeAId8?: number;
	cancelOtherFormCritOnSuccess?: boolean;
	cancelOtherFormCritOnSuccessRanking?: number;
	lastChangeUser?: string;
	lastChangeDate?: string;
	contProd?: UtPcaContProdId;
	utPcaFormDefnId?: Array<UtPcaFormDefnId>;
	constructor(obj: any = {}) {
		this.formId = obj.formId;
		this.contProdId = obj.contProdId;
		this.deletedFlag = obj.deletedFlag;
		this.formNm = obj.formNm;
		this.formCritCntId = obj.formCritCntId;
		this.formTypeAId = obj.formTypeAId;
		this.formTypeBId = obj.formTypeBId;
		this.formTypeCId = obj.formTypeCId;
		this.mktbsktId = obj.mktbsktId;
		this.mktbsktId1 = obj.mktbsktId1;
		this.mktbsktId2 = obj.mktbsktId2;
		this.mktbsktId3 = obj.mktbsktId3;
		this.mktbsktId4 = obj.mktbsktId4;
		this.mktbsktId5 = obj.mktbsktId5;
		this.mktbsktId6 = obj.mktbsktId6;
		this.mktbsktId7 = obj.mktbsktId7;
		this.mktbsktId8 = obj.mktbsktId8;
		this.formNote = obj.formNote;
		this.clinicalEditEval = obj.clinicalEditEval;
		this.formCreateDate = obj.formCreateDate;
		this.formUpdateDate = obj.formUpdateDate;
		this.formTypeAId1 = obj.formTypeAId1;
		this.formTypeAId2 = obj.formTypeAId2;
		this.formTypeAId3 = obj.formTypeAId3;
		this.formTypeAId4 = obj.formTypeAId4;
		this.formTypeAId5 = obj.formTypeAId5;
		this.formTypeAId6 = obj.formTypeAId6;
		this.formTypeAId7 = obj.formTypeAId7;
		this.formTypeAId8 = obj.formTypeAId8;
		this.cancelOtherFormCritOnSuccess = obj.cancelOtherFormCritOnSuccess;
		this.cancelOtherFormCritOnSuccessRanking = obj.cancelOtherFormCritOnSuccessRanking;
		this.lastChangeUser = obj.lastChangeUser;
		this.lastChangeDate = obj.lastChangeDate;
		this.contProd = new UtPcaContProdId(obj.contProd);
		this.utPcaFormDefnId = obj.utPcaFormDefnId.map(o => new UtPcaFormDefnId(o));
	}
}